cloudapi: split tests into logical groupings
